Fix duplicate / missing graph
be02e697772ddd6cd6998f0b87aeb6d8448b0ecf added a duplicate entry of
    Western Europe, Daily Increase in Confirmed Cases Per Million Inhabitants
    <img src="27mar2020/rates.png">
at this spot in the file.  Based on the table of contents, this should
have been the rate increases for the Nordic region.

The .png file was also missing from the git repository; I generated it
on my system, but I don't have the Helvetica font so it looks different.
That should be trivial to fix by running "make" again on your system.